Moved files to algos/ directory for organization
